fix(api): Mapped[Decimal] sur colonnes DECIMAL + garde state-machine sur assign_ticket
Audit qualité — findings MEDIUM :
- Modèles : 36 colonnes DECIMAL(15,2)/(5,2)/(10,2) annotées Mapped[float] alors que
  SQLAlchemy/asyncpg renvoie des Decimal au runtime. Passage en Mapped[Decimal] (+ import).
  Aligne le typage strict sur la convention « montants en Decimal » et évite qu'un dev
  fasse de l'arithmétique float par erreur. mypy reste vert (0 ripple : le runtime était
  déjà en Decimal). overall_score (Float, pas DECIMAL) laissé en float.
- maintenance.assign_ticket : forçait status="assigned" sans passer par la machine à
  états → un ticket in_progress/resolved/closed/cancelled pouvait être ressuscité. Ajout
  d'une garde is_valid_transition (ré-assignation assigned→assigned toujours autorisée).

ruff + format(app) + mypy(187) OK.

+++ Desktop/SGI/SGI_Project/apps/api/app/models/finance.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import DECIMAL, Date, DateTime, ForeignKey, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID
@@ -29,7 +30,7 @@ class FinanceTransaction(Base, TimestampMixin, TenantMixin, SoftDeleteMixin):
     direction: Mapped[str] = mapped_column(String(10), nullable=False, default="debit")
 
     # Montant
-    amount: Mapped[float] = mapped_column(DECIMAL(15, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(DECIMAL(15, 2), nullable=False)
     currency: Mapped[str] = mapped_column(String(3), nullable=False, default="AED")
 
     # Statut
